Downloads/vibe coding/ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
openai==1.6.1
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0
pymongo==4.6.0
boto3==1.34.0
pyyaml==6.0.1
pytest==7.4.3
pytest-mock==3.12.0
